Emit signal specific exit code (#4001)
* Emit signal specific exit code

* Add changelog

* Handle process termination signals with specific exit codes when CLI is not active
