+++ .changeset/twenty-trams-fetch.md
@@ -1,0 +1,5 @@
+---
+"@kilocode/cli": minor
+---
+
+Support emitting Unix exit codes for signal interruption (SIGINT/SIGTERM).
